Vss2Git

+++ Source/OldCode/Archive/2007 Phasor Protocols/PhasorProtocols/IDigitalValue.cs
@@ -25,6 +25,6 @@ namespace PCS.PhasorProtocols
         /// <summary>
         /// Gets or sets the 16-bit integer value (composed of digital bits) that represents this <see cref="IDigitalValue"/>.
         /// </summary>
-        int Value { get; set; }
+        short Value { get; set; }
     }
 }
